Create all_code
This is all the code I have plus the analysis. Part of the analysis I combined with "analysis of code"  from your file. Also, this whole thing will be easier to look at when u knit it to an html file. Please feel free to add more or edit or anything.